pyproject.toml
[tool.poetry]
name = "airlines-crawlers"
version = "0.1.0"
description = ""
authors = ["Cardcurator Team"]
readme = "README.md"
packages = [{include = "airlines_crawlers"}]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.15.2"
undetected-chromedriver = "^3.5.3"
retry = "^0.9.2"
bs4 = "^0.0.1"
lxml = "^4.9.3"
python-dotenv = "^1.0.0"
mysql-connector-python = "^8.3.0"
curl-cffi = "^0.6.2"
schedule = "^1.2.2"


[tool.poetry.group.dev.dependencies]
flake9 = "^3.8.3.post2"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 119
extend-ignore = "E203,"
exclude = "*migrations*"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
